+++ tests/utilities/test_executors.py
@@ -8,7 +8,8 @@ from unittest.mock import MagicMock
 import pytest
 
 import prefect
-from prefect.utilities.executors import Heartbeat, timeout_handler
+from prefect.utilities.configuration import set_temporary_config
+from prefect.utilities.executors import Heartbeat, timeout_handler, run_with_heartbeat
 
 
 def test_heartbeat_calls_function_on_interval():
@@ -26,6 +27,79 @@ def test_heartbeat_calls_function_on_interval():
     timer.cancel()
     timer.join()
     assert a.called == 2
+
+
+def test_heartbeat_logs_if_first_call_fails(caplog):
+    class A:
+        def __init__(self):
+            self.logger = prefect.utilities.logging.get_logger("A")
+
+        def _heartbeat(self):
+            raise SyntaxError("oops")
+
+        @run_with_heartbeat
+        def run(self):
+            pass
+
+    a = A()
+    a.run()
+
+    assert caplog.records
+
+    log = caplog.records[0]
+    assert log.name == "prefect.A"
+    assert "Heartbeat" in log.message
+    assert "zombie" in log.message
+
+
+def test_heartbeat_logs_if_first_call_fails(caplog):
+    class A:
+        def __init__(self):
+            self.logger = prefect.utilities.logging.get_logger("A")
+
+        def _heartbeat(self):
+            raise SyntaxError("oops")
+
+        @run_with_heartbeat
+        def run(self):
+            pass
+
+    a = A()
+    a.run()
+
+    assert caplog.records
+
+    log = caplog.records[0]
+    assert log.name == "prefect.A"
+    assert "Heartbeat" in log.message
+    assert "zombie" in log.message
+
+
+def test_heartbeat_logs_if_thread_dies(caplog):
+    class A:
+        def __init__(self):
+            self.calls = 0
+            self.logger = prefect.utilities.logging.get_logger("A")
+
+        def _heartbeat(self):
+            if self.calls == 1:
+                raise SyntaxError("oops")
+
+        @run_with_heartbeat
+        def run(self):
+            self.calls = 1
+            time.sleep(1)
+
+    a = A()
+    with set_temporary_config({"cloud.heartbeat_interval": 0.25}):
+        a.run()
+
+    assert caplog.records
+
+    log = caplog.records[0]
+    assert log.name == "prefect.A"
+    assert "Heartbeat thread died" in log.message
+    assert "zombie" in log.message
 
 
 def test_timeout_handler_times_out():